requirementx.txt
asgiref==3.5.2
deprecation==2.1.0
dj-database-url==1.0.0
Django==4.0.6
django-environ==0.9.0
django-on-heroku==1.1.2
django-social-share==2.3.0
gunicorn==20.1.0
img2pdf==0.4.4
lxml==4.9.1
packaging==21.3
pikepdf==5.4.2
Pillow==9.2.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
qrcode==7.3.1
sqlparse==0.4.2
whitenoise==6.2.0
